add tests for all signal sources
