+++ asset_test.go
@@ -54,59 +54,3 @@ func TestAssetSink(t *testing.T) {
 		assertEqual(t, "samples", asset.Signal.Length(), test.samples)
 	}
 }
-
-func TestSource(t *testing.T) {
-	sampleData := [][]float64{{0, 1, 2, 3, 4, 5, 6, 7, 8, 9}}
-	buf := signal.Allocator{
-		Channels: len(sampleData),
-		Length:   len(sampleData[0]),
-		Capacity: len(sampleData[0]),
-	}.Float64()
-	signal.WriteStripedFloat64(sampleData, buf)
-	sampleRate := signal.SampleRate(44100)
-
-	tests := []struct {
-		source   pipe.SourceAllocatorFunc
-		expected []float64
-		msg      string
-	}{
-		{
-			source:   audio.Source(sampleRate, buf),
-			expected: []float64{0, 1, 2, 3, 4, 5, 6, 7, 8, 9},
-			msg:      "Full asset",
-		},
-		{
-			source:   audio.Source(sampleRate, buf.Slice(0, 3)),
-			expected: []float64{0, 1, 2},
-			msg:      "First three",
-		},
-		{
-			source:   audio.Source(sampleRate, buf.Slice(1, 3)),
-			expected: []float64{1, 2},
-			msg:      "Two from within",
-		},
-		{
-			source:   audio.Source(sampleRate, buf.Slice(5, 10)),
-			expected: []float64{5, 6, 7, 8, 9},
-			msg:      "Last five",
-		},
-	}
-
-	bufferSize := 2
-	for _, test := range tests {
-		sink := mock.Sink{}
-		l, _ := pipe.Routing{
-			Source: test.source,
-			Sink:   sink.Sink(),
-		}.Line(bufferSize)
-
-		p := pipe.New(context.Background(), pipe.WithLines(l))
-		_ = p.Wait()
-
-		result := make([]float64, sink.Values.Len())
-		signal.ReadFloat64(sink.Values, result)
-
-		assertEqual(t, test.msg, result, test.expected)
-	}
-
-}
